changed the iterator interface into abstract class
